Q2 Complete - Circle.py
Coded. Plotted. Added image to resource folder of plotting
====================================================
Succint Description for TeX file :
The first line of the text file represents the number of vertices and faces. Based on that we load up the vertices as Point Objects into a dictionary and then faces into a dictionary, as Face objects which contain those Point objects. The curve to be generated is to be defined using the quadratic basis therefore the Face class contains those basis functions as b0, b1, and b2.

After loading up all faces, for all $0\leq t \leq 1$, with each $t$ set at a difference of $0.0001$, we calculate the Bezier curve for each face by,

\textbf{}\begin{center}
    $P(t) = (1-t)^2 \boldsymbol{P}_0 + 2(1-t)(t)\boldsymbol{P}_1 + t^2\boldsymbol{P}_2$ \cite{NADNotes14}
\end{center}{}

We plot the control points and the generated Bezier curves which results in figure \ref{fig:BezierbasisQ2}

\begin{figure}
\includegraphics[scale=1]{resources/Q2_QBezBasis.png}
\centering
\caption{Figure Generated using the control points and quadratic Bezier basis functions}
\label{fig:BezierbasisQ2}
\end{figure}

\textbf{(see Appendix 2 for Code)}